+++ tests/common.py
@@ -1,3 +1,4 @@
+import inspect
 import re
 import shutil
 import subprocess
@@ -17,8 +18,11 @@ class Dashboard:
     port: int = attr.ib()
     process: subprocess.Popen = attr.ib()
 
-    def state(self):
-        r = requests.get(f"http://localhost:{self.port}/api/state", timeout=10)
+    def state(self, *, id=None):
+        r = requests.get(
+            f"http://localhost:{self.port}/api/state/{'' if id is None else id}",
+            timeout=10,
+        )
         r.raise_for_status()
         return r.json()
 
@@ -35,22 +39,25 @@ def wait_for(condition, *, timeout, interval):
 
 
 @contextmanager
-def dashboard(*, port: int = 0, test_dir: Optional[Path] = None) -> Dashboard:
+def dashboard(*, port: int = 0, test_path: Optional[Path] = None) -> Dashboard:
     """
     Launches a dashboard process with --dashboard-only. Defaults to a random open
     port.
     """
     args = ["hypothesis", "fuzz", "--dashboard-only", "--port", str(port)]
-    if test_dir is not None:
-        args += ["--", str(test_dir)]
+    if test_path is not None:
+        args += ["--", str(test_path)]
     process = subprocess.Popen(args, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
     port = None
     # wait for dashboard to start up
     for _ in range(100):
         time.sleep(0.05)
         if process.poll() is not None:
+            stdout = b"".join(process.stdout.readlines())
+            stderr = b"".join(process.stderr.readlines())
             raise Exception(
-                f"dashboard invocation exited with return code {process.returncode}"
+                f"dashboard invocation exited with return code {process.returncode}. "
+                f"stdout:\n{stdout}\nstderr:\n{stderr}"
             )
 
         # flask prints the port on stderr :(
@@ -75,12 +82,15 @@ def dashboard(*, port: int = 0, test_dir: Optional[Path] = None) -> Dashboard:
         dashboard.process.stderr.close()
         dashboard.process.kill()
         dashboard.process.wait()
-        if test_dir is not None:
-            shutil.rmtree(test_dir)
+        if test_path is not None:
+            if test_path.is_dir():
+                shutil.rmtree(test_path)
+            else:
+                test_path.unlink()
 
 
 @contextmanager
-def fuzz(*, n=1, dashboard=False, test_dir=None):
+def fuzz(*, n=1, dashboard=False, test_path=None):
     args = [
         "hypothesis",
         "fuzz",
@@ -88,8 +98,8 @@ def fuzz(*, n=1, dashboard=False, test_dir=None):
         str(n),
         "--dashboard" if dashboard else "--no-dashboard",
     ]
-    if test_dir is not None:
-        args += ["--", str(test_dir)]
+    if test_path is not None:
+        args += ["--", str(test_path)]
     process = subprocess.Popen(args)
 
     try:
@@ -103,18 +113,13 @@ BASIC_TEST_CODE = """
 import json
 import time
 
-from hypothesis import given, settings, strategies as st
-from hypothesis.database import DirectoryBasedExampleDatabase
-
-settings.register_profile("testing", settings(database=DirectoryBasedExampleDatabase("{}")))
-settings.load_profile("testing")
-
 # some non-trivial test that will exercise a bunch of lines
 jsons = st.deferred(lambda: st.none() | st.integers() | st.floats() | st.text() | lists | objects)
 lists = st.lists(jsons)
 objects = st.dictionaries(st.text(), jsons)
 
 def to_jsonable(obj):
+    time.sleep(0.05)
     if isinstance(obj, int):
         # create a bunch of artificial branches
         if abs(obj) <= 100:
@@ -161,7 +166,6 @@ def to_jsonable(obj):
     if isinstance(obj, dict):
         return {{str(k): to_jsonable(v) for k, v in obj.items()}}
 
-    time.sleep(0.1)
     return str(obj)
 
 @given(jsons)
@@ -170,9 +174,24 @@ def test(x):
 """
 
 
-def write_basic_test_code():
+def write_test_code(code):
     db_dir = Path(tempfile.mkdtemp())
     test_dir = Path(tempfile.mkdtemp())
+
+    code = (
+        inspect.cleandoc(
+            f"""
+            from hypothesis import given, settings, strategies as st
+            from hypothesis.database import DirectoryBasedExampleDatabase
+            import pytest
+
+            settings.register_profile("testing", settings(database=DirectoryBasedExampleDatabase("{db_dir}")))
+            settings.load_profile("testing")
+            """
+        )
+        + "\n"
+        + inspect.cleandoc(code)
+    )
     test_file = test_dir / "test_a.py"
-    test_file.write_text(BASIC_TEST_CODE.format(str(db_dir)))
+    test_file.write_text(code)
     return (test_dir, db_dir)